+++ R/finalfit_internal_functions.R
@@ -719,7 +719,7 @@ globalVariables(c("L95", "U95", "fit_id", "Total",
 									"OR", "HR", "Coefficient", ".", ".id", "var", "value",
 									":=", "Mean", "SD", "Median", "Q3", "Q1", "IQR", "Formatted", 
 									"w", "Freq", "g", "total_prop", "Prop", "index_total", "vname", "Combined",
-									"2.5 %", "97.5 %", "p.value", "estimate", "index"))
+									"2.5 %", "97.5 %", "p.value", "estimate", "index", "n"))
 
 
 # Workaround ::: as summary.formula not (yet) exported from Hmisc