+++ R/summaryfactorlist.R
@@ -69,7 +69,7 @@
 
 summary_factorlist <- function(.data, dependent = NULL, explanatory, cont = "mean", cont_cut = 5,
 															 p = FALSE, digits = c(1, 1, 3, 1), na_include = FALSE,
-															 column = FALSE, total_col = FALSE, orderbytotal = FALSE, fit_id = FALSE,
+															 column = TRUE, total_col = FALSE, orderbytotal = FALSE, fit_id = FALSE,
 															 na_to_missing = TRUE, add_dependent_label = FALSE,
 															 dependent_label_prefix = "Dependent: ", dependent_label_suffix = "", ...) {
 	if(!is.data.frame(.data)) stop(".data is not dataframe")
@@ -246,7 +246,7 @@ summary_factorlist0 <- function(.data, dependent, explanatory,  cont = "mean", c
 #' @keywords internal
 summary_factorlist_groups <- function(.data, dependent, explanatory,  cont = "mean", cont_cut = 5, 
 																			p = FALSE, digits = c(1, 1, 3, 1), na_include = FALSE,
-																			column = FALSE, total_col = FALSE, orderbytotal = FALSE, fit_id = FALSE,
+																			column = TRUE, total_col = FALSE, orderbytotal = FALSE, fit_id = FALSE,
 																			na_to_missing = TRUE, add_dependent_label = FALSE,
 																			dependent_label_prefix = "Dependent: ", dependent_label_suffix = "", ...){